[*] Build fix

+++ modules/common/include/dap_chain_datum_tx_receipt.h
@@ -34,8 +34,8 @@
 typedef struct dap_chain_datum_tx_receipt {
     dap_chain_tx_item_type_t type; // Transaction item type
     dap_chain_receipt_info_t receipt_info; // Receipt itself
-    uint16_t size;
-    uint16_t exts_size;
+    uint64_t size;
+    uint64_t exts_size;
     byte_t exts_n_signs[]; // Signatures, first from provider, second from client
 }DAP_ALIGN_PACKED dap_chain_datum_tx_receipt_t;
